Pipeline can now send output to provided output directory
Default behaviour remains the same as before. Note, there's a chance
that a seg fault will occur if BISCUIT v1.0.0 or earlier is used -
so use v.1.0.1 if possible. Note, the only files that won't be sent
to the specified directory are the two log files from
bin/run_snakemake_workflow.sh, which will go to the logs directory
at the same level as the Snakefile.