Big changes in the pipeline
Rules have been split off into reasonably grouped rules files and are
included using include in Snakefile. Pipeline updated to only need to
run once to create all desired files (no need to run once for renaming
and once for everything else). Updated the submit script to create the
logs/runs directory for those output files and to point to new config
directory. Added preseq to config, as well as changing NULL to "NULL"
so that it will work properly if uncommented.