Add conda envs, percent genome covered plot, and change trimming options
Conda envs is now the default method for which packages to use. Percent
genome covered plot will always be produced since it uses BISCUITqc
output and is quick to generate. Trimming options is now given as a list
in the config.yaml file. Specify the trim galore options and values you
want here.